Add simple UI field structures and list. Rename UI data drawing functions, reuse code in field rendering functions.

+++ portapack_hackrf/firmware/portapack_hackrf/lcd_loop.h
@@ -26,9 +26,9 @@
 
 typedef struct device_state_t {
 	int64_t tuned_hz;
-	int8_t lna_gain_db;
-	int8_t if_gain_db;
-	int8_t bb_gain_db;
+	int32_t lna_gain_db;
+	int32_t if_gain_db;
+	int32_t bb_gain_db;
 
 	uint32_t duration_decimate;
 	uint32_t duration_channel_filter;
